Move remaining hidden tests to proper tests dir.
