+++ lib/stsci/tools/mputil.py
@@ -1,7 +1,10 @@
 #!/usr/bin/env python
 #
-from __future__ import print_function
-import math, time
+from __future__ import division, print_function
+
+import math
+import time
+
 
 class WatchedProcess(object):
     """ MINIMAL wrapper around multiprocessing.Process so we can more easily track/time them. """
@@ -52,7 +55,6 @@ def launch_and_wait(mp_proc_list, pool_size):
     procs = []
     for p in mp_proc_list:
         procs.append(WatchedProcess(p))
-    num_total = len(procs)
 
     # Launch all of them, but only so pool_size are running at any time
     keep_going = True
@@ -100,43 +102,6 @@ def launch_and_wait(mp_proc_list, pool_size):
             raise RuntimeError("Problem during: "+str(p.process.name)+ \
                   ', exitcode: '+str(p.process.exitcode)+'. Check log.')
     # all is well, can return
-
-
-def takes_time(x):
-    """ Example function which takes some time to run - just here for testing. """
-
-    import numpy
-    START = time.time()
-    s = numpy.float64(1)
-    #s = numpy.float32(1)
-    #s = 1.0
-
-    assert x not in (3, 7,9), "Simulate some errors"
-
-    for i in range(10000000):
-        s = (s + x) * s % 2399232
-
-    elap = time.time() - START
-    print(('Done "takes_time" x='+str(x)+': s = '+str(s)+', elapsed time = %.2f s' % elap))
-
-
-def do_main():
-    """ Illustrate use of launch_and_wait """
-    # load em up
-    import multiprocessing
-    p = None
-    subprocs = []
-    for item in [2,3,4,5,6,7,8,9]:
-        print(("mputil: instantiating Process for x = "+str(item)))
-        p = multiprocessing.Process(target=takes_time, args=(item,),
-                                    name='takes_time()')
-        subprocs.append(p)
-
-    # launch em, pool-fashion
-    launch_and_wait(subprocs, 3)
-
-    # by now, all should be finished
-    print("All subprocs should be finished and joined.")
 
 
 def best_tile_layout(pool_size):
@@ -193,21 +158,3 @@ def best_tile_layout(pool_size):
     xnum = int(math.sqrt(pool_size))
     ynum = int((1.*pool_size)/xnum)
     return (xnum, ynum)
-
-
-def test_best_tile_layout():
-    """ Loop though some numbers and make sure we get expected results. """
-    for i in range(257):
-        x,y = best_tile_layout(i)
-        assert (x*y <= i) or (i == 0), "Total num resulting tiles > pool_size"
-        unused_cores = i - (x*y)
-        print(i, (x,y), unused_cores)
-        if i < 10:
-            assert unused_cores <= 1, "Too many idle cores at i = "+str(i)
-        else:
-            percent_unused = 100.*((unused_cores*1.)/i)
-            assert percent_unused < 14., "Too many idles cores at i: "+str(i)
-
-
-if __name__=='__main__':
-    do_main()
