+++ lib/stsci/tools/tests/test_tkrotext.py
@@ -1,0 +1,45 @@
+from __future__ import absolute_import
+
+import sys
+import time
+
+from ..tkrotext import ROText
+
+PY3K = sys.version_info[0] > 2
+
+if PY3K:
+    import tkinter as TKNTR
+else:
+    import Tkinter as TKNTR
+
+
+def test_rotest():
+    rot = None
+
+    def quit():
+        sys.exit()
+
+    def clicked():
+        rot.insert(TKNTR.END, "\nClicked at " + time.asctime(), force=True)
+        rot.see(TKNTR.END)
+
+    # make our test window
+    top = TKNTR.Tk()
+    f = TKNTR.Frame(top)
+
+    sc = TKNTR.Scrollbar(f)
+    sc.pack(side=TKNTR.RIGHT, fill=TKNTR.Y)
+    rot = ROText(f, wrap=TKNTR.WORD, height=10, yscrollcommand=sc.set,
+                 focusBackTo=top)
+    rot.pack(side=TKNTR.TOP, fill=TKNTR.X, expand=True)
+    sc.config(command=rot.yview)
+    f.pack(side=TKNTR.TOP, fill=TKNTR.X)
+
+    b = TKNTR.Button(top, text='Click Me', command=clicked)
+    b.pack(side=TKNTR.TOP, fill=TKNTR.X, expand=1)
+
+    q = TKNTR.Button(top, text='Quit', command=quit)
+    q.pack(side=TKNTR.TOP)
+
+    # start
+    top.mainloop()
